add ruff and precommit (#7)
A personal deficiency of mine is that I'm very easily distracted by
things that can be automated away by tools. I hope it won't be too much
overhead to have these hooks installed. It feels better to do this
earlier rather than later so that we guarantee a lack of "format
everything" PRs in the future (at least in Python).