add pyright CI (#8)
For similar reasons to #7, I'd love to not have the type checking errors
in the CI when I open the project. More importantly, xDSL actually is
invalid if you pass wrong types in to places, so downstream projects
kind of need to be conformant to the pyright rules at least at the
boundary, so it feels prudent to have this check.

---------

Co-authored-by: hpompougnac <hpompougnac@inria-Latitude-5520.inrialpes.fr>